Revert "lock journal before reading journal files"
This reverts commit 47358a6f956681346a198cdf27c60bcf9498cc45.

This added overhead, and will not be needed, because appends are going
to have to be made atomic for other reasons than avoiding incomplete
reads of data being appended.

In particular, when git-annex is interrupted in the middle of an append,
it must not leave the file with a partially written line. So appending
has to somehow be made fully atomic.