remove explicit-setup-deps
This has not been needed since stack <1.4.0, and even the i386ancent
build uses stack 2.1.1.

Stack 2.7.5 seems to have forgotten about this old config and warns
about it, so this avoids that warning.

The libtinfo-dev was added to the docs at the same time, I assume it is
also not necessary.