Revert "stack.yaml: Updated to lts-19.16"
This reverts commit b5dc04099efd8b1bd2dbfa09a288518eabf8ceec.

Broke windows build, because the new lts updates Win32 to a version that
lacks a function that git-annex needs. git-annex.cabal depends on an
older Win32, and so stack build fails.

Will need to wait to update stack.yaml until this is fixed
https://github.com/haskell/win32/issues/208
and is in a new LTS release.