[DATALAD RUNCMD] Double the timeout to 3600 for git-annex test
We had OSX build failing recently due to exit 124 indicative of timeout.
It might be that git-annex got slower (ping @joeyh just in case) but
also possible that more tests now make the run slower.  To not mess with
benchmarking etc, decided to just double the timeout threshold over the board
and 3600 was already timeout on some clients

=== Do not change lines below ===
{
 "chain": [],
 "cmd": "git-sedi 'timeout 1800' 'timeout 3600'",
 "exit": 0,
 "extra_inputs": [],
 "inputs": [],
 "outputs": [],
 "pwd": "."
}
^^^ Do not change lines above ^^^